requirements.txt
asgiref==3.3.4
asyncpg==0.22.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
databases==0.4.3
Django==3.2
django-request==1.5.6
ecdsa==0.14.1
fastapi==0.63.0
gunicorn==20.1.0
h11==0.12.0
passlib==1.7.4
psycopg2==2.8.6
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2021.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.24
sqlparse==0.4.1
starlette==0.13.6
typing-extensions==3.7.4.3
unicorn==1.0.2
uvicorn==0.13.4
websockets==8.1
